Enforce the invariant that populating HTML assets is blocked against.
The intended behaviour when dealing with HTML assets is to include
population part of the promise which is blocked against by other
parts of the code e.g. client notification on path change. Prior
to these changes this was not guaranteed based on which asset load
function was used. Address this by making loadAsset the canonical
method to call and have it choose the correct function internally.