Make use of rollup node-resolve code for ESM node module resolution.
In order to grab the correct path to to the ESM file of a package
the correct fields we need to be able to decode potentially multiple
keys within the package.json file and also understand the "exports"
object. Rather than duplicate this logic, make use of code within a
rollup plugin that knows how to resolve these in node.