Remove the call to loadAsset when a "none" relevance asset is served.
Take the decision that we no longer need to place this asset in the
graph because it isn't possible to know if as leaf for it will ever
be resolved (e.g. it could be loaded asynchronously by JS code). It
happens that always reload mode the server will still kick clients
for such assets if they change thus remove it without issue.