Rework import resolution to contain the logic centrally.
Convert the assetMiddleware so it is always "dumb" and accepts only
relative paths within the node_modules directory for static modules.

Move all the logic for converting from an npm installed module ref
into an ImportResolver that is instantiated only once and called
from both the PathMonitor and the assetMiddleware on module serving.