Add support for monorepos.
Detect relivestyle being executed in circumstances where it is serving
out a public folder of a sub-project located within a monorepo - this
can be managed both as yarn workspaces and via lerna - and if so allow
resolving modules symlinked to packages.

Achieve this by allowing relative paths being supplied to requests for
__node_modules. Make use of this to serve packages within the monorepo
that may be symlinked to its counterparts.