Fix resolution of imports within symlinked modules.
The module resolution code borrowed from rollup must have its rootDir
set to within the symlinked package path in order for modules required
by a package itself to be correctly resolved. Arrange this taking care
that the returned paths, as supplied to the client, are paths relative
to the original root directory i.e. the path being served.