Consolidate resolution of hoisted modules & contain in ImportResolver.
Detect the case of hosited node_modules by checking whether the resolved
module directory is different from the nearest node_modules path. If so,
construct and return a relative path from resolve() in the same way as
is done for adjacent packages in monorepos.

This allows the removal of all module resolution code and monorepo
awareness from the asset middleware.