Bump rollup module resolution to 11.2.1 & fix tests via ~htm version.
Be conservative around pulling in updated module resolution code and
do a first jump to the latest of the version 11 series. During module
re-installation it was also noted that htm has adjusted the naming of
its browser facing modules, so address the test failures by forcing
use of the last working version via a tilde version. This ensures we
can adopt changes more slowly and thus validate them more thoroughly.