Ensure rewriting consumes the raw text of a JavaScript asset.
This fixes an issue where the parse + generate done internally by
assetgraph when changed code text is assigned to a JavaScript asset,
which is done to allow outgoing relations to be recalculated, causes
a comment handling bug in escodegen to be triggered and can lead to
invalid code (linebreaks being added before return statements). Use
a facility added in assetgraph 7.1.0 to keep the raw text available
as the _text property and always use this as the basis for serving.