Upgrade rollup module resolution to 13.0.4 which fixes self-require.
It seemed that naming a file being imported the same as the module
it was importing would cause some misbehaviour causing resolution
to return the file itself rather than the target module.

On a hunch that this might have been addressed by newer versions of
the resolution code pull in a newer version and integrate it - this
gets what was an important thing out of the way while also getting
rid of that issue. Add a test to cover the specific situation.